rename ntqq->qqbot
