+++ assets/config_store/zh_cn/bot_qqbot.toml
@@ -2,7 +2,7 @@
 # 注意：TOML 不是 Python。请不要在此处使用 Python 语法。
 # 例如：TOML 中的布尔值必须是小写。
 
-[bot_ntqq]
+[bot_qqbot]
 # 平台端的配置部分，此处填写的值可在消息中以明文形式展示。请不要在此部分填写敏感信息。
 qq_bot_appid = "<Replace me with (int, str) value>" # QQ 机器人（官方）的 AppID。
 enable = false # 是否启用此平台。
@@ -10,6 +10,6 @@ qq_private_bot = false # 使用 QQ 机器人（官方）时，是否为私域机
 qq_bot_enable_send_url = false # 使用 QQ 机器人（官方）时，是否启用发送 URL。
 qq_typing_emoji = 181 # 使用 QQ 相关协议端时，正在处理消息时向消息挂上的回应表情 ID，需协议端支持。ID 详见：https://bot.q.qq.com/wiki/develop/api/openapi/emoji/model.html#Emoji
 
-[bot_ntqq_secret]
+[bot_qqbot_secret]
 # 平台端的密钥配置部分，此处的值若意外出现在发送的消息中，机器人会尝试拦截。但请务必提防泄露。
 qq_bot_secret = "<Replace me with str value>" # QQ 机器人（官方）的 AppSecret。
