+++ assets/config_store/zh_tw/bot_qqbot.toml
@@ -2,7 +2,7 @@
 # 注意：TOML 不是 Python。請不要在此處使用 Python 語法。
 # 例如：TOML 中的布林值必須是小寫。
 
-[bot_ntqq]
+[bot_qqbot]
 # 平台端的設定部分，此處填寫的值可在訊息中以明文顯示。請不要在此部分填寫敏感資訊。
 qq_bot_appid = "<Replace me with (int, str) value>" # QQ 機器人（官方）的 AppID。
 enable = false # 是否啟用此平台。
@@ -10,6 +10,6 @@ qq_private_bot = false # 使用 QQ 機器人（官方）時，是否為私域機
 qq_bot_enable_send_url = false # 使用 QQ 機器人（官方）時，是否啟用傳送 URL。
 qq_typing_emoji = 181 # 使用 QQ 相關協定端時，正在處理訊息時向訊息掛上的回應表情 ID，需協定端支援。 ID 詳見：https://bot.q.qq.com/wiki/develop/api/openapi/emoji/model.html#Emoji
 
-[bot_ntqq_secret]
+[bot_qqbot_secret]
 # 平台端的金鑰設定部分，此處的值若意外出現在傳送的訊息中，機器人會嘗試攔截。但請務必提防洩漏。
 qq_bot_secret = "<Replace me with str value>" # QQ 機器人（官方）的 AppSecret。
