+++ bots/qqbot/bot.py
@@ -5,8 +5,8 @@ import sys
 import botpy
 from botpy.message import C2CMessage, DirectMessage, GroupMessage, Message
 
-from bots.ntqq.info import *
-from bots.ntqq.message import MessageSession, FetchTarget
+from bots.qqbot.info import *
+from bots.qqbot.message import MessageSession, FetchTarget
 from core.bot_init import init_async, load_prompt
 from core.builtins import PrivateAssets, Url
 from core.config import Config
@@ -15,11 +15,11 @@ from core.constants.info import Info
 from core.parser.message import parser
 from core.types import MsgInfo, Session
 
-PrivateAssets.set(os.path.join(assets_path, 'private', 'ntqq'))
+PrivateAssets.set(os.path.join(assets_path, 'private', 'qqbot'))
 Info.dirty_word_check = Config('enable_dirty_check', False)
 Url.disable_mm = False
-qq_appid = str(Config("qq_bot_appid", cfg_type=(int, str), table_name='bot_ntqq'))
-qq_secret = Config("qq_bot_secret", cfg_type=str, secret=True, table_name='bot_ntqq')
+qq_appid = str(Config("qq_bot_appid", cfg_type=(int, str), table_name='bot_qqbot'))
+qq_secret = Config("qq_bot_secret", cfg_type=str, secret=True, table_name='bot_qqbot')
 
 
 class MyClient(botpy.Client):
@@ -159,12 +159,12 @@ class MyClient(botpy.Client):
         await parser(msg, prefix=prefix, require_enable_modules=require_enable_modules)
 
 
-if Config("enable", False, table_name='bot_ntqq'):
+if Config("enable", False, table_name='bot_qqbot'):
     intents = botpy.Intents.none()
     intents.public_guild_messages = True
     intents.public_messages = True
     intents.direct_message = True
-    if Config('qq_private_bot', False, table_name='bot_ntqq'):
+    if Config('qq_private_bot', False, table_name='bot_qqbot'):
         intents.guild_messages = True
 
     client = MyClient(intents=intents, bot_log=None)
